backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0
apscheduler==3.10.4
aiohttp==3.9.1
beautifulsoup4==4.12.2
feedparser==6.0.10
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
